# Add optimized method for broacasting ismissing