--- a/src/array.jl
+++ b/src/array.jl
@@ -790,3 +790,7 @@ refs(A::CategoricalArray) = A.refs
 pool(A::CategoricalArray) = A.pool
 
 Base.deleteat!(A::CategoricalArray, inds) = (deleteat!(A.refs, inds); A)
+
+Base.Broadcast.broadcasted(::typeof(ismissing), A::CategoricalArray{T}) where {T} =
+    T >: Missing ? Base.Broadcast.broadcasted(==, A.refs, 0) :
+                   Base.Broadcast.broadcasted(_ -> false, A)