This will allow StatsModel to get rid of the dependency on CategoricalArrays (https://github.com/JuliaStats/StatsModels.jl/pull/157), and more generally provide much improved performance for this common operation.

Also clean some uses of internal `SubArray` fields.

Cc: @kleinschmidt

```julia
using BenchmarkTools, CategoricalArrays

x = categorical(rand([1:10; missing], 10_000));
nonmissing!(mask, x) = mask .|= ismissing.(x)
mask = similar(x, Bool);

# Before
julia> @btime ismissing.(x);
  36.188 μs (5 allocations: 5.58 KiB)

julia> @btime nonmissing!(mask, x);
  25.366 μs (0 allocations: 0 bytes)

# After
julia> @btime ismissing.(x);
  15.578 μs (5 allocations: 5.59 KiB)

julia> @btime nonmissing!(mask, x);
  1.723 μs (0 allocations: 0 bytes)